+++ src/test/java/com/hhplus/concert_ticketing/application/concert/ConcertConcurrencyTest.java
@@ -4,10 +4,10 @@ import com.hhplus.concert_ticketing.application.ConcertUseCase;
 import com.hhplus.concert_ticketing.domain.concert.*;
 import com.hhplus.concert_ticketing.domain.queue.Queue;
 import com.hhplus.concert_ticketing.domain.queue.QueueStatus;
-import com.hhplus.concert_ticketing.infra.concert.JpaConcertRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaPerformanceRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaSeatRepository;
-import com.hhplus.concert_ticketing.infra.queue.JpaQueueRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaConcertRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaPerformanceRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaSeatRepository;
+import com.hhplus.concert_ticketing.infra.repository.queue.JpaQueueRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
