+++ src/test/java/com/hhplus/concert_ticketing/application/concert/ConcertUseCaseIntegrationTest.java
@@ -7,12 +7,12 @@ import com.hhplus.concert_ticketing.domain.queue.Queue;
 import com.hhplus.concert_ticketing.domain.queue.QueueStatus;
 import com.hhplus.concert_ticketing.domain.user.Point;
 import com.hhplus.concert_ticketing.domain.user.Users;
-import com.hhplus.concert_ticketing.infra.concert.JpaConcertRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaPerformanceRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaSeatRepository;
-import com.hhplus.concert_ticketing.infra.queue.JpaQueueRepository;
-import com.hhplus.concert_ticketing.infra.user.JpaPointRepository;
-import com.hhplus.concert_ticketing.infra.user.JpaUsersRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaConcertRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaPerformanceRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaSeatRepository;
+import com.hhplus.concert_ticketing.infra.repository.queue.JpaQueueRepository;
+import com.hhplus.concert_ticketing.infra.repository.user.JpaPointRepository;
+import com.hhplus.concert_ticketing.infra.repository.user.JpaUsersRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
