+++ src/test/java/com/hhplus/concert_ticketing/application/concert/PaymentConcurrencyTest.java
@@ -4,18 +4,16 @@ import com.hhplus.concert_ticketing.application.ConcertUseCase;
 import com.hhplus.concert_ticketing.domain.concert.*;
 import com.hhplus.concert_ticketing.domain.queue.Queue;
 import com.hhplus.concert_ticketing.domain.queue.QueueStatus;
-import com.hhplus.concert_ticketing.infra.concert.JpaConcertRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaPerformanceRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaReservationRepository;
-import com.hhplus.concert_ticketing.infra.concert.JpaSeatRepository;
-import com.hhplus.concert_ticketing.infra.queue.JpaQueueRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaConcertRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaPerformanceRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaReservationRepository;
+import com.hhplus.concert_ticketing.infra.repository.concert.JpaSeatRepository;
+import com.hhplus.concert_ticketing.infra.repository.queue.JpaQueueRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
